\name{big.brother}
\alias{big.brother}

\title{Verbosely Track Function Evaluations}

\description{This is for making function evaluations more verbose;
  \code{big.brother} returns a function that prints its arguments and
  return value every iteration (or every several iterations if
  \code{interval} is specified)}

\usage{
big.brother(f, interval=1)
}

\arguments{
  \item{f}{A function}
  \item{interval}{Interval between printed evaluations}
}

\keyword{programming}
