\name{check}
\alias{check.fftC}

\title{Check Capabilities of the Diversitree Install}

\description{These check to see if FFTW support was included in
  diversitree.  They rarely need to be called directly.}

\usage{
check.fftC(error=TRUE)
}

\arguments{
  \item{error}{Logical: causes an error if FFTW is not
    available if \code{TRUE}}
}

\author{Richard G. FitzJohn}
\keyword{utilities}
