\name{diversitree-package}
\alias{diversitree-package}
\alias{diversitree}
\docType{package}
\title{
\packageTitle{diversitree}
}
\description{
\packageDescription{diversitree}
}
\author{
\packageAuthor{diversitree}

Maintainer: \packageMaintainer{diversitree}
}
\references{
  Diversitree contains methods described in the following papers (all of
  which aside from Maddison et al. 2007 were originally published as a
  diversitree implementation).

  \itemize{
    \item{FitzJohn R.G., Maddison W.P., and Otto S.P. 2009. Estimating
      trait-dependent speciation and extinction rates from incompletely
      resolved phylogenies. systematic biology 58:595-611. Systematic
      Biology 58:595-611.}

    \item{FitzJohn R.G. 2010. Quantitative traits and
      diversification. Systematic Biology 59:619-633.}

    \item{Goldberg E.E., Lancaster L.T., Ree R.H. 2011. Phylogenetic
    inference of reciprocal effects between geographic range evolution
    and diversification.  Systematic Biology 60: 451-465.}

    \item{Maddison W.P., Midford P.E., and Otto S.P. 2007.  Estimating a
      binary character's effect on speciation and extinction.
      Systematic Biology 56: 701-710.}

    \item{Magnuson-Ford K. and Otto S.P. 2012. Linking the
      investigations of character evolution and species
      diversification. The American Naturalist 180: 225-245.}
  }
}
\keyword{ package }
