\name{geosse.phy}
\docType{data}
\alias{geosse.phy}
\title{Example GeoSSE Tree}
\description{
  This is a test GeoSSE tree, created by Emma Goldberg
}
\usage{geosse.phy}
\format{An 'ape' tree, with additional element \code{tip.state} that
  includes the tip states.}
\source{Generated from a C program by Emma Goldberg}
\keyword{datasets}