\name{Constants}
\alias{ROOT.FLAT}
\alias{ROOT.EQUI}
\alias{ROOT.OBS}
\alias{ROOT.GIVEN}
\alias{ROOT.BOTH}
\alias{ROOT.ALL}

\title{Constants Used by Diversitree}

\usage{
ROOT.FLAT
ROOT.EQUI
ROOT.OBS
ROOT.GIVEN
ROOT.BOTH
ROOT.ALL
}

\description{
  These constants are used by \code{\link{make.bisse}}, 
  \code{\link{make.geosse}} and \code{\link{make.mkn}}
  for specifying root behaviour.
}

\keyword{internal}
